WIP: Make initial state checks work with shortcuts.
