+++ Packages/com.unity.inputsystem/InputSystem/InputManagerStateMonitors.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 1407189163374c4aa71b9689243809eb
+timeCreated: 1647431896
